Move `router-act.ts` to `e2e-utils` and delete outdated duplicate
In #84521, The test helper `router-act.ts` was copied from the
`segment-cache` test suite to the `app-prefetch` test suite. This led to
some confusion as the two copies diverged in #84603. To avoid this
problem, the recently updated file has now been moved to `e2e-utils`,
and the now outdated duplicate has been deleted.